Extract Codex Connector review-loop prompt sections
